pyproject.toml
[project]
name = "crewdev"
version = "0.1.0"
description = "crewdev using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.157.0,<1.0.0",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.1.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0"
]

[project.scripts]
crewdev = "crewdev.main:run"
run_crew = "crewdev.main:run"
train = "crewdev.main:train"
replay = "crewdev.main:replay"
test = "crewdev.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
